Separate classification models and statistical evaluation scripts
